moved code around and created modal component
