+++ src/pages/index.js
@@ -1,2 +1,2 @@
-export * from "./Home";
-export * from "./Help";
+export * from './Home';
+export * from './Help/Help';
